requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
sqlite3